fix main type cli args

+++ WARP/main.py
@@ -15,9 +15,9 @@ def parse_args():
     parser.add_argument('--I', type=int)
     parser.add_argument('--M', type=int)
     parser.add_argument('--T', type=int)
-    parser.add_argument('--nu', type=int)
-    parser.add_argument('--mu', type=int)
-    parser.add_argument('--lambda', type=int)
+    parser.add_argument('--nu', type=float)
+    parser.add_argument('--mu', type=float)
+    parser.add_argument('--lambda', type=float)
     return parser.parse_args()
 
 def load_config(json_file):
